refactor: finish 2 step

+++ server/src/main/java/com/example/distributed_system/service/DistributorServiceImpl.java
@@ -1,8 +1,9 @@
 package com.example.distributed_system.service;
 
+import com.example.distributed_system.dto.DistributorCreateDto;
 import com.example.distributed_system.entity.DistributionCommittee;
-import com.example.distributed_system.entity.Distributor;
 import com.example.distributed_system.entity.World;
+import com.example.distributed_system.interfaces.DistributorService;
 import com.example.distributed_system.repository.DistributorRepository;
 import com.example.distributed_system.repository.WorldRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,19 +14,21 @@ import java.util.Comparator;
 
 @Service
 @Transactional
-public class DistributorService {
+public class DistributorServiceImpl implements DistributorService {
     @Autowired
     private DistributorRepository distributorRepository;
     @Autowired
     private WorldRepository worldRepository;
 
+    @Override
     public void deleteDistributor(Integer distributorID) {
-//        мне в падлу делать проверку)
-        distributorRepository.deleteById(distributorID);
+        if (distributorRepository.existsById(distributorID))
+            distributorRepository.deleteById(distributorID);
     }
 
-    public void addDistributor(Integer worldId, Distributor distributor) {
-        World world = worldRepository.findById(worldId).orElseThrow();
+    @Override
+    public void addDistributor(DistributorCreateDto distributor) {
+        World world = worldRepository.findById(distributor.getWorldId()).orElseThrow();
         DistributionCommittee committee = world.getDistributionLayer().getDistributionCommittees().stream()
                 .min(Comparator.comparing(distributionCommittee -> distributionCommittee.getDistributors().size())).orElseThrow();
         if (committee.getDistributors().size() < 5) {
